chore: types for watch mode
