+++ package.json
@@ -28,9 +28,11 @@
     "prepare": "is-ci || husky install"
   },
   "devDependencies": {
+    "@microsoft/api-extractor": "7.31.2",
     "@sindresorhus/slugify": "1.1.2",
     "@slack/webhook": "6.1.0",
     "@types/benchmark": "2.1.2",
+    "@types/fs-extra": "9.0.13",
     "@types/glob": "8.0.0",
     "@types/graphviz": "0.0.34",
     "@types/node": "14.18.32",
@@ -55,6 +57,7 @@
     "eslint-plugin-simple-import-sort": "8.0.0",
     "eventemitter3": "4.0.7",
     "execa": "5.1.1",
+    "fs-extra": "10.1.0",
     "glob": "8.0.3",
     "globby": "11.1.0",
     "graphviz-mit": "0.0.9",
