+++ tsconfig.build.regular.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "incremental": true,
     "target": "es2018",
     "module": "commonjs",
     "lib": ["es2018", "ES2021.WeakRef"],
